[FIX] area import queue jobs field
