+++ spp_base_common/models/__init__.py
@@ -4,3 +4,4 @@ from . import phone_validation
 from . import phone_number
 from . import queue_job
 from . import res_partner
+from . import area_import
