+++ spp_base_common/views/area_import_view.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="view_spparea_import_form_inherit" model="ir.ui.view">
+        <field name="name">view_spparea_import_form_inherit</field>
+        <field name="model">spp.area.import</field>
+        <field name="inherit_id" ref="spp_area_base.view_spparea_import_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//page[@name='json_data']" position="after">
+                <page string="Queue Jobs" name="queue_jobs" invisible="not job_ids">
+                    <field name="job_ids" nolabel="1" readonly="1">
+                        <tree>
+                            <field name="name" />
+                            <field
+                                name="state"
+                                decoration-info="state=='pending'"
+                                decoration-warning="state=='enqueued'"
+                                decoration-primary="state=='started'"
+                                decoration-success="state=='done'"
+                                decoration-danger="state=='failed'"
+                                widget="badge"
+                            />
+                            <field name="date_created" />
+                            <field name="date_started" />
+                            <field name="date_done" />
+                            <field name="exc_info" />
+                        </tree>
+                        <form>
+                            <header>
+                                <field name="state" widget="statusbar" />
+                            </header>
+                            <group>
+                                <group>
+                                    <field name="name" />
+                                    <field name="model_name" />
+                                    <field name="method_name" />
+                                    <field name="res_model" />
+                                    <field name="res_id" />
+                                </group>
+                                <group>
+                                    <field name="date_created" />
+                                    <field name="date_enqueued" />
+                                    <field name="date_started" />
+                                    <field name="date_done" />
+                                    <field name="priority" />
+                                </group>
+                            </group>
+                            <group string="Exception Information" invisible="not exc_info">
+                                <field name="exc_info" nolabel="1" readonly="1" />
+                            </group>
+                        </form>
+                    </field>
+                </page>
+            </xpath>
+        </field>
+    </record>
+</odoo>
